change fund to pool, get auth data from req.user object
