+++ src/models/index.js
@@ -2,4 +2,4 @@ require('./Account')
 require('./Game')
 require('./Prize')
 require('./User')
-require('./Fund')
+require('./Pool')
